fix(adsense): duplicate `beforeInteractive` scripts appended to <head>
Culprit: NextJS bug during SSR where `beforeInteractive` scripts
(even with a unique `key` specified) were appended to <head> for each
time the component was rendered. The `/news/:author` pages were most
affected, as `infeed` ads being placed after every 5 articles had the
effect of adding the same `adsbygoogle.js` script as many times.

NextJS docs state that `beforeInteractive` scripts can't be used outside
`_document.tsx`, even though it worked in the sense that it placed them
in the <head> tag, as we wanted: https://nextjs.org/docs/messages/no-before-interactive-script-outside-document.

Not sure if this duplication behaviour was a regression, or if it was always a "bug".